Move constant elements to Layout
