+++ src/layouts/index.js
@@ -1,6 +1,22 @@
+import styles from "./Layout.module.css"
+
+
+const TitleSection = () => {
+    return <div className={styles.mediamingleParent}>
+        <b className={styles.mediamingle}>MediaMingle</b>
+        <b className={styles.anAppThat}>
+            An app that lets you search for music, tweets AND movies!
+        </b>
+    </div>
+}
+
+
 const Layout = ({ children }) => {
     return <div>
-        {children}
+        <section className={styles.home}>
+            <TitleSection />
+            {children}
+        </section>
     </div>
 }
