--- src/pages/Home.module.css
@@ -1,51 +1,0 @@
-.anAppThat,
-.mediamingle {
-  position: absolute;
-  margin: 0 !important;
-  top: 0;
-  left: 2px;
-  display: inline-block;
-  width: 511px;
-  height: 43px;
-  flex-shrink: 0;
-  z-index: 0;
-}
-.anAppThat {
-  top: 57px;
-  left: 0;
-  font-size: var(--font-size-xl);
-  width: 784px;
-  height: 103px;
-  z-index: 1;
-}
-.home,
-.mediamingleParent {
-  display: flex;
-  flex-direction: row;
-  box-sizing: border-box;
-  align-items: flex-start;
-  justify-content: flex-start;
-}
-.mediamingleParent {
-  margin: 0 !important;
-  position: absolute;
-  top: 103px;
-  left: 70px;
-  height: 94px;
-  padding: 10px;
-  gap: 10px;
-  z-index: 3;
-}
-.home {
-  position: relative;
-  background-color: var(--color-white);
-  width: 100%;
-  height: 832px;
-  overflow: hidden;
-  padding: 40px;
-  gap: 40px;
-  text-align: left;
-  font-size: 50px;
-  color: var(--color-black);
-  font-family: var(--font-inter);
-}
